requirements.txt
autopep8==1.6.0
pycodestyle==2.8.0
python-dotenv==0.20.0
toml==0.10.2
watchdog==2.1.7
